Ensure expansion of leftover pending nodes
Certain transports require pathing through areas of very unfavourable heuristic (usually caves very far away) to get to the path target, with the result that these transports are never expanded. This is solved by expanding any leftover pending nodes when the ordinary expansion (non-transports) is finished. However, this could lead to the shortest path not necessarily using the transport with the shortest travel time to the target, although I am yet to find an example of this in practice.