Improve heuristic for transport nodes
Rather than forcing all leftover pending nodes to be expanded we improve the heuristic of the transport nodes to ensure that expansion of transport nodes will happen. The new heuristic for transport nodes will use the best heurisitc of the origin and destination node. Transports that are direct shortcuts will remain the same and use the destination node to calculate the heuristic, while transports with considerable detours will now use the origin node instead.