Update collision map to 2023-07-26 and remove header
Updated the collision map to revision 215 from the 2023-07-26 DT2 release.
Removed the 16 byte minX, minY, maxX, maxY collision map region header, as this can easily be determined from the file name, e.g. `50_50` becomes `minX = 50 * 64 = 3200` and `maxX = minX + 64 - 1`. This was done because the size/layout of the collision map was already partially hardcoded and it saves roughly `16 * 1500 = 24000` bytes of data when downloading the plugin.